# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API document for UAMS backend server",
        "title": "Backend API",
        "contact": {
            "name": "Mr. Khai"
        },
        "version": "1.0"
    },
    "host": "http://iot.hcmue.space:8079",
    "basePath": "/v1",
    "paths": {
...
        "/v1/gateway_logs": {
            "get": {
                "description": "find all gateway logs info",
                "produces": [
                    "application/json"
                ],
                "summary": "Find All GatewayLog",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.GatewayLog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/gateway_id/{gateway_id}/period/{from}/{to}": {
            "get": {
                "description": "find Gateway logs by period of time",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Gateway logs by period of time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Unix time",
                        "name": "from",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To Unix time",
                        "name": "to",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GatewayLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/gateway_id/{id}": {
            "get": {
                "description": "find gateway log info by Gateway id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find GatewayLog By Gateway ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GatewayLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/period/{from}/{to}": {
            "get": {
                "description": "find Gateway logs by period of time",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Gateway logs by period of time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Unix time",
                        "name": "from",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To Unix time",
                        "name": "to",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GatewayLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/{id}": {
            "get": {
                "description": "find gateway log info by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find GatewayLog By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GatewayLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.GatewayLog": {
            "type": "object",
            "properties": {
                "gateway_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "log_time": {
                    "type": "string"
                },
                "state_type": {
                    "type": "string"
                },
                "state_value": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMsg": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}